worker/requirements.txt
redis==5.0.8
python-json-logger==2.0.7
dkimpy==1.1.0